Fix: Resolve type error on signup page
