+++ peer_web_frontend/src/app/signup/panel/SignUpField.tsx
@@ -1,40 +1,10 @@
 'use client'
-import { Controller, Control, FieldError } from 'react-hook-form'
-import { Typography, Button } from '@mui/material'
-import CuTextField from '@/components/CuTextField'
+import { Controller } from 'react-hook-form'
+import { Typography, Button, TextField } from '@mui/material'
+// import CuTextField from '@/components/CuTextField'
 import CuTextFieldLabel from '@/components/CuTextFieldLabel'
 
-import { ISignUpInputs } from '@/types/ISignUpInputs'
-
-interface ISignUpFieldRules {
-  required: string
-  pattern?: {
-    value: RegExp
-    message: string
-  }
-  minLength?: {
-    value: number
-    message: string
-  }
-  maxLength?: {
-    value: number
-    message: string
-  }
-  validate?: (value: string) => boolean | string
-}
-
-interface ISignUpField {
-  label: string
-  name: string
-  control: Control<ISignUpInputs, any>
-  error: FieldError
-  rules: ISignUpFieldRules
-  placeholder?: string
-  onClick?: () => void
-  buttonText?: string
-  isInputValid?: boolean
-  inputProps?: any
-}
+import { ISignUpField } from '@/types/ISignUpInputs'
 
 const SignUpField = ({
   label,
@@ -57,7 +27,7 @@ const SignUpField = ({
         render={({ field }) => (
           <>
             <CuTextFieldLabel htmlFor={name}>{label}</CuTextFieldLabel>
-            <CuTextField
+            <TextField
               {...field}
               type="text"
               disabled={isInputValid}
